+++ packages/backend-proxy-middleware-cf/CHANGELOG.md
@@ -1,5 +1,11 @@
 # @sap-ux/backend-proxy-middleware-cf
 
+## 0.0.79
+
+### Patch Changes
+
+- @sap-ux/adp-tooling@0.18.89
+
 ## 0.0.78
 
 ### Patch Changes