+++ packages/repo-app-import-sub-generator/CHANGELOG.md
@@ -1,5 +1,34 @@
 # @sap-ux/repo-app-download-sub-generator
 
+## 0.3.280
+
+### Patch Changes
+
+- @sap-ux/fiori-elements-writer@2.8.97
+- @sap-ux/odata-service-inquirer@2.19.11
+
+## 0.3.279
+
+### Patch Changes
+
+- @sap-ux/fiori-elements-writer@2.8.96
+- @sap-ux/odata-service-inquirer@2.19.11
+
+## 0.3.278
+
+### Patch Changes
+
+- Updated dependencies [32d028b]
+    - @sap-ux/odata-service-inquirer@2.19.11
+
+## 0.3.277
+
+### Patch Changes
+
+- @sap-ux/fiori-elements-writer@2.8.95
+- @sap-ux/inquirer-common@0.11.23
+- @sap-ux/odata-service-inquirer@2.19.10
+
 ## 0.3.276
 
 ### Patch Changes